src/main/webapp/app/entities/provice/delete/provice-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ProviceService } from '../service/provice.service';

import { ProviceDeleteDialogComponent } from './provice-delete-dialog.component';

describe('Provice Management Delete Component', () => {
  let comp: ProviceDeleteDialogComponent;
  let fixture: ComponentFixture<ProviceDeleteDialogComponent>;
  let service: ProviceService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, ProviceDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(ProviceDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ProviceDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ProviceService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
